avoid initial indexing if the index directory is present

+++ scripts/start.sh
@@ -1,34 +1,35 @@
 #!/bin/bash
 
-# Reindex default
+# default period for reindexing (in minutes)
 if [ -z "$REINDEX" ]; then
 	REINDEX=10
 fi
 
-#START METHOD FOR INDEXING OF OPENGROK
-start_opengrok(){
+indexer(){
 	# Wait for Tomcat startup.
-	date +"%F %T Waiting for tomcat startup..."
+	date +"%F %T Waiting for Tomcat startup..."
 	while [ "`curl --silent --write-out '%{response_code}' -o /dev/null 'http://localhost:8080/'`" == "000" ]; do
 		sleep 1;
 	done
 	date +"%F %T Startup finished"
 
-	# Populate the webapp with bare configuration.
-	BODY_INCLUDE_FILE="/data/body_include"
-	if [[ -f $BODY_INCLUDE_FILE ]]; then
-		mv "$BODY_INCLUDE_FILE" "$BODY_INCLUDE_FILE.orig"
-	fi
-	echo '<p><h1>Waiting on the initial reindex to finish.. Stay tuned !</h1></p>' > "$BODY_INCLUDE_FILE"
-	/scripts/index.sh --noIndex
-	rm -f "$BODY_INCLUDE_FILE"
-	if [[ -f $BODY_INCLUDE_FILE.orig ]]; then
-		mv "$BODY_INCLUDE_FILE.orig" "$BODY_INCLUDE_FILE"
-	fi
+	if [[ ! -d /data/index ]]; then
+		# Populate the webapp with bare configuration.
+		BODY_INCLUDE_FILE="/data/body_include"
+		if [[ -f $BODY_INCLUDE_FILE ]]; then
+			mv "$BODY_INCLUDE_FILE" "$BODY_INCLUDE_FILE.orig"
+		fi
+		echo '<p><h1>Waiting on the initial reindex to finish.. Stay tuned !</h1></p>' > "$BODY_INCLUDE_FILE"
+		/scripts/index.sh --noIndex
+		rm -f "$BODY_INCLUDE_FILE"
+		if [[ -f $BODY_INCLUDE_FILE.orig ]]; then
+			mv "$BODY_INCLUDE_FILE.orig" "$BODY_INCLUDE_FILE"
+		fi
 
-	# Perform initial indexing.
-	/scripts/index.sh
-	date +"%F %T Initial reindex finished"
+		# Perform initial indexing.
+		/scripts/index.sh
+		date +"%F %T Initial reindex finished"
+	fi
 
 	# Continue to index every $REINDEX minutes.
 	if [ "$REINDEX" == "0" ]; then
@@ -43,6 +44,6 @@ start_opengrok(){
 	done
 }
 
-#START ALL NECESSARY SERVICES.
-start_opengrok &
+# Start all necessary services.
+indexer &
 catalina.sh run
